19:52-8/30-基础功能基本完成

+++ src/main/java/com/sy/zy/controller/RoleController.java
@@ -2,9 +2,11 @@ package com.sy.zy.controller;
 
 import com.github.pagehelper.PageInfo;
 import com.sy.bmq.model.Role;
+import com.sy.bmq.model.User;
 import com.sy.bmq.model.base.BaseResult;
 import com.sy.zy.service.RoleService;
 import com.sy.zy.service.UserService;
+import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -21,6 +23,8 @@ public class RoleController {
     UserService userService;
 
 
+    //查询所有集合分页
+    @RequiresPermissions("/role/selectall.do")
     @RequestMapping("/selectall.do")
     public BaseResult findAll(Integer page, Integer limit) throws Exception {
         BaseResult baseResult = new BaseResult();
@@ -38,17 +42,27 @@ public class RoleController {
         }
     }
 
+    //删除角色，关联用户表，存在就不删除
     @RequestMapping("/delete.do")
     public BaseResult deletebyId(int id) throws Exception {
         BaseResult baseResult = new BaseResult();
-        Integer delete = roleService.delete(id);
-        if (delete > 0) {
-            baseResult.setCode(0);
-            return baseResult;
-        } else {
-            baseResult.setCode(1);
+        User user=new User();
+        user.setRoleId(id);
+        List<User> users = userService.selectAll(user);
+        if(users.size()==0){
+            Integer delete = roleService.delete(id);
+            if (delete > 0) {
+                baseResult.setCode(0);
+                return baseResult;
+            } else {
+                baseResult.setCode(1);
+                return baseResult;
+            }
+        }else {
+            baseResult.setMsg("failed");
             return baseResult;
         }
+
     }
 
     @RequestMapping("/update.do")
@@ -80,13 +94,16 @@ public class RoleController {
     }
 
 
+
     @RequestMapping("/selectcount.do")
     public BaseResult selectCount(String roleName) throws Exception {
         BaseResult baseResult = new BaseResult();
-        Integer integer = userService.selectCount(roleName);
-        System.out.println("asdasdasdasdasdsadsadasdasdasdasdsdsad"+integer);
-        baseResult.setCount(integer);
-        baseResult.setCode(0);
+        if(roleName!=("管理员")){
+            Integer integer = userService.selectCount(roleName);
+            baseResult.setCount(integer);
+            baseResult.setData(roleName);
+            baseResult.setCode(0);
+        }
         return baseResult;
     }
